Move main files to core/
