+++ DarkChess/core/log.cpp
@@ -2,7 +2,7 @@
 #include "spdlog/spdlog.h"
 #include "spdlog/sinks/stdout_color_sinks.h"
 
-#include "DarkChess/log.hpp"
+#include "DarkChess/core/log.hpp"
 
 namespace DarkChess
 {
@@ -12,7 +12,7 @@ std::shared_ptr<spdlog::logger> Log::s_client_logger;
 
 void Log::init()
 {
-    spdlog::set_pattern("%^[%T] (%l) %n: %v%$");
+    spdlog::set_pattern("[%T] %^(%n %l) %v%$");
 
     s_core_logger = spdlog::stdout_color_mt("DarkChess Engine");
     s_core_logger->set_level(spdlog::level::trace);
