+++ DarkChess/core/log.hpp
@@ -28,10 +28,3 @@ public:
 #define DC_CORE_WARN(...) ::DarkChess::Log::get_core_logger()->warn(__VA_ARGS__)
 #define DC_CORE_ERROR(...) ::DarkChess::Log::get_core_logger()->error(__VA_ARGS__)
 #define DC_CORE_FATAL(...) ::DarkChess::Log::get_core_logger()->fatal(__VA_ARGS__)
-
-// Client Macros
-#define DC_TRACE(...) ::DarkChess::Log::get_client_logger()->trace(__VA_ARGS__)
-#define DC_INFO(...) ::DarkChess::Log::get_client_logger()->info(__VA_ARGS__)
-#define DC_WARN(...) ::DarkChess::Log::get_client_logger()->warn(__VA_ARGS__)
-#define DC_ERROR(...) ::DarkChess::Log::get_client_logger()->error(__VA_ARGS__)
-#define DC_FATAL(...) ::DarkChess::Log::get_client_logger()->fatal(__VA_ARGS__)
